BudgetGenerator improvements

+++ budget-generator/src/main/java/org/scouthub/budgetgenerator/application/DeleteActivity.java
@@ -1,16 +1,11 @@
 package org.scouthub.budgetgenerator.application;
 
 import org.scouthub.budgetgenerator.domain.repository.ActivityRepository;
-import org.scouthub.budgetgenerator.domain.service.BudgetService;
 
 public final class DeleteActivity {
   private DeleteActivity() {}
 
-  public static void delete(
-      Long activityId, ActivityRepository activityRepository, BudgetService budgetService) {
-    if (activityRepository.existsById(activityId)) {
-      activityRepository.deleteById(activityId);
-      budgetService.delete(activityId);
-    }
+  public static void delete(Long activityId, ActivityRepository activityRepository) {
+    if (activityRepository.existsById(activityId)) activityRepository.deleteById(activityId);
   }
 }
